Increase SRTM I2C Service buffer length
The default buffer length for the SRTM I2C Service (16 bytes) is too small to
accomodate commands for some I2C devices. This was due to a limitation of the
RPMsg I2C driver in Linux which restricts the max I2C message size to the size
of the SRTM I2C Service buffer length. This has been increased on the Linux side
(to 128 bytes), and the same has been done here on the MCU side.